WFP-998 changed conviction id to big integer

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/client/dto/Conviction.kt
@@ -1,12 +1,13 @@
 package uk.gov.justice.digital.hmpps.hmppsworkload.client.dto
 
 import com.fasterxml.jackson.annotation.JsonCreator
+import java.math.BigInteger
 import java.time.LocalDateTime
 
 data class Conviction @JsonCreator constructor(
   val sentence: Sentence?,
   val custody: Custody?,
-  val convictionId: Long,
+  val convictionId: BigInteger,
   val courtAppearance: CourtAppearance?,
   val offences: List<Offence>?
 )
