+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsworkload/mapper/CaseTypeMapper.kt
@@ -3,12 +3,13 @@ package uk.gov.justice.digital.hmpps.hmppsworkload.mapper
 import org.springframework.stereotype.Service
 import uk.gov.justice.digital.hmpps.hmppsworkload.client.dto.Conviction
 import uk.gov.justice.digital.hmpps.hmppsworkload.domain.CaseType
+import java.math.BigInteger
 
 @Service
 class CaseTypeMapper(
   private val caseTypeRules: List<CaseTypeRule>
 ) {
-  fun getCaseType(activeConvictions: List<Conviction>, convictionId: Long): CaseType {
+  fun getCaseType(activeConvictions: List<Conviction>, convictionId: BigInteger): CaseType {
     val allConvictionTypes = activeConvictions.map { conviction ->
       convictionToCaseType(conviction).let { caseType ->
         {
